refactor: extract TTS playback into shared tts module
Remove duplicate `speak_via_headless` and `split_tts_segments`
implementations
from nc_router and sq_router. Also fix headless_channel initialization
to only
require `headless.tts.enabled` instead of also checking `always_tts`.
